+++ CHANGELOG.md
@@ -16,12 +16,13 @@ Whenever `minor` is bumped in the `major.minor.patch` version (e.g. `1.0.2 -> 1.
   - Various frontend bugfixes
   - Allow for campaign files to contain multiple campaigns in a list
   - Better time tracking
-  - Allow loading an existing campagin with `add-existing`
+  - Add loading of an existing campagin with `add-existing`
   - Add a button for downloading campaign sources
   - Move campaign source data storate from `data/tasks/` to `data/campaigns/` with automatic migration
   - Add global progress tracking to the dashboard
   - Restyle dashboard
   - Add image zoom on click
+  - Add baking of annotations into a static frontend with `bake-existing`
 - v1.1.4
   - Make span toolboxes stick for 0.5 second after selection
 - v1.1.3